1. More fixes to file prefixes support 2. Added Gentoo Linux ebuild to simplify install process 3. Warehouse terrain maps updated to the latest version from J-F Pruneau
git-svn-id: https://ufo2000.svn.sourceforge.net/svnroot/ufo2000/trunk@286 80a73507-52ba-0310-bdc1-b9b4a1856fab

+++ init-scripts/filecheck.lua
@@ -3,6 +3,7 @@ local function X(a) return tonumber(a, 16) end
 FilesTable = {
 	["$(ufo2000)/ufo2000.dat"]  = { Crc32 = X("0x8EA4C104") },
 	["$(ufo2000)/keyboard.dat"] = { Crc32 = X("0xC796755E") },
+	["$(ufo2000)/soundmap.xml"] = { Crc32 = X("0x00000000") },
 
 	["$(home)/geodata.dat"] = {
 		Origin = "$(ufo2000)/geodata.dat" },
@@ -155,8 +156,20 @@ local function CheckSingleDataFile(name, info)
 			if fh == nil then
 				return string.format("CheckDataFiles: can't open '%s' for read/write", name)
 			end
-			fh:close()
-			os.remove(name)
+
+			if info.Origin then
+				local data = ReadFile(LocateFile(info.Origin))
+				if data == nil then
+					return string.format(
+						"CheckDataFiles: can't get original file '%s' for initialization", 
+						info.Origin) 
+				end
+				fh:write(data)
+				fh:close()
+			else
+				fh:close()
+				os.remove(name)
+			end
 		else
 			fh:close()
 		end
